+++ src/main/scala/jp/riken/massbank/reader/scala/types/Types.scala
@@ -14,9 +14,9 @@ trait SingleLine extends LineType
 trait MultipleLine extends LineType
 
 trait Peak
-case class CompletePeakTriple(mz: Int, int: Double, rel: Double) extends Peak
-case class AbsolutePeakPair(mz: Int, int: Double) extends Peak
-case class RelativePeakPair(mz: Int, rel: Double) extends Peak
+case class CompletePeakTriple(mz: Double, int: Double, rel: Double) extends Peak
+case class AbsolutePeakPair(mz: Double, int: Double) extends Peak
+case class RelativePeakPair(mz: Double, rel: Double) extends Peak
 
 case class PeakData(peaks: List[Peak]) {
   val hasRelativeIntensity: Boolean = false
